Fixes Analysis Querry
Fixes teh querry used to populate the RHTargets table with the values of Nu corresponding to the best fits. Before we picked an arbitrary value of 0.05 to be within Q-0.5 now we select a Nu > 100 and select the closest Q to 0.5. This removes the issue of getting null values due to a dataset that has Q values that cross 0.5 but never are within +/- 0.05 of it.