Adds SQLTestHelper.
The SQLTestHelper is used to setup the test MySQL database with
the 'chamber' schema (database). The database is created when an
instance of the helper is created. So far the only methods
implemented include run_script and clear_db.

The only script that exists is createdb.sql. Next steps include
adding an experiment_test_data.sql script to add some data into the
tables.