Reformates the `DataSpec` contract.
The `DataSpec` contract not uses a `pool_id` rather than a `PoolSpec`
instance. This is because the Tdms file only knows the pool_id.

The code sends a message to the user that a pool must be added if
it doesn't already exist.

Also, the `add_data` test was moved from unit testing into integration
testing to remove the need for mocks. As a result, all of the unit
tests for `ChamberAccess` have been removed.